host functions, pass by state version enum

+++ frame/executive/src/lib.rs
@@ -671,7 +671,7 @@ mod tests {
 
 			#[pallet::weight(0)]
 			pub fn calculate_storage_root(_origin: OriginFor<T>) -> DispatchResult {
-				let root = sp_io::storage::root(sp_runtime::StateVersion::V1 as u8);
+				let root = sp_io::storage::root(sp_runtime::StateVersion::V1);
 				sp_io::storage::set("storage_root".as_bytes(), &root);
 				Ok(())
 			}
