+++ frame/support/src/lib.rs
@@ -751,9 +751,9 @@ macro_rules! assert_noop {
 		$x:expr,
 		$y:expr $(,)?
 	) => {
-		let h = $crate::storage_root($crate::StateVersion::V1 as u8);
+		let h = $crate::storage_root($crate::StateVersion::V1);
 		$crate::assert_err!($x, $y);
-		assert_eq!(h, $crate::storage_root($crate::StateVersion::V1 as u8));
+		assert_eq!(h, $crate::storage_root($crate::StateVersion::V1));
 	};
 }
 
@@ -766,9 +766,9 @@ macro_rules! assert_storage_noop {
 	(
 		$x:expr
 	) => {
-		let h = $crate::storage_root($crate::StateVersion::V1 as u8);
+		let h = $crate::storage_root($crate::StateVersion::V1);
 		$x;
-		assert_eq!(h, $crate::storage_root($crate::StateVersion::V1 as u8));
+		assert_eq!(h, $crate::storage_root($crate::StateVersion::V1));
 	};
 }
