+++ frame/support/src/storage/child.rs
@@ -171,7 +171,7 @@ pub fn put_raw(child_info: &ChildInfo, key: &[u8], value: &[u8]) {
 pub fn root(child_info: &ChildInfo, version: sp_core::StateVersion) -> Vec<u8> {
 	match child_info.child_type() {
 		ChildType::ParentKeyId =>
-			sp_io::default_child_storage::root(child_info.storage_key(), version as u8),
+			sp_io::default_child_storage::root(child_info.storage_key(), version),
 	}
 }
